requirements.txt
asgiref==3.7.2
branca==0.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==2023.10
crispy-tailwind==0.5.0
decorator==5.1.1
distlib==0.3.7
Django==4.2.3
django-betterforms==2.0.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==23.5
django-use-email-as-username==1.4.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.2
folium==0.14.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
idna==3.4
Jinja2==3.1.2
Markdown==3.5.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
numpy==1.25.0
Pillow==9.5.0
platformdirs==3.9.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytweening==1.0.7
pytz==2023.3.post1
ratelim==0.1.6
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sslcommerz-lib==1.0
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.24.2
